+++ CHANGELOG.md
@@ -8,8 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added/Changed/Fixed/Removed  
 - Nothing  
 
+## [1.0.1] - 2020-10-17  
+### Fixed  
+- Additional folder was created even when output option had a folder selected
+- If output is a filename, it must end with .zip  
+
 ## [1.0.0] - 2020-10-16  
 - Initial release  
 
-[Unreleased]: https://github.com/diegozanon/get-github-code/compare/v1.0.0...HEAD  
+[Unreleased]: https://github.com/diegozanon/get-github-code/compare/v1.0.1...HEAD  
+[1.0.1]: https://github.com/diegozanon/get-github-code/releases/tag/v1.0.1  
 [1.0.0]: https://github.com/diegozanon/get-github-code/releases/tag/v1.0.0